src/requirements.txt
numpy
h5py
tqdm
keras>=2.3.0
tensorflow>=2.2
opencv-python
tqdm
mlflow
scikit-learn
matplotlib
scikit-image
streamlit
psutil
memory-profiler
pytorch-gradcam
torch
